[clean] コンポーネントを分割
コンポーネントを分割し、コードの可読性を上げた。
