+++ pages/index.js
@@ -1,69 +1,18 @@
 import React, { useReducer, useState } from 'react'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import reducer from './reducers/events'
-import Event from './components/Event'
+import Events from './components/Events'
+import EventForm from './components/EventForm'
 
 
 const App = () => {
   const [state, dispatch] = useReducer(reducer, []);
-  const [title, setTitle] = useState("")
-  const [body, setBody] = useState("")
-
-
-  const addEvent = e => {
-    e.preventDefault()
-    dispatch({
-      type: "CREATE_EVENT",
-      title,
-      body
-    })
-    setTitle("")
-    setBody("")
-  }
-
-  const unCreate = title === "" || body === ""
-
-
-  const deleteAllEvents = e => {
-    e.preventDefault()
-    const result = window.confirm("全てのイベントを本当に削除しても良いですか？")
-    if (result) dispatch({type: "DELETE_ALL_EVENTS"})
-  }
-
 
   return (
     <>
       <div className="container-fluid">
-        <h4>イベントフォーム作成</h4>
-        <form>
-          <div className="form-group">
-            <label htmlFor="formEventTitle">タイトル</label>
-            <input className="form-control" id="formEventTitle" value={title} onChange={e => setTitle(e.target.value)}></input>
-          </div>
-
-          <div className="form-group">
-            <label htmlFor="formEventBody">ボディー</label>
-            <input className="form-control" id="formEventBody" value={body} onChange={e => setBody(e.target.value)}></input>
-          </div>
-
-          <button className="btn btn-primary" onClick={addEvent} disabled={unCreate}>イベントを作成する</button>
-          <button className="btn btn-danger" onClick={deleteAllEvents} disabled={state.length === 0}>全てのイベントを削除する</button>
-        </form>
-
-        <h4>イベント一覧</h4>
-        <table className="table table-hover">
-          <thead>
-            <tr>
-              <th>ID</th>
-              <th>タイトル</th>
-              <th>ボディー</th>
-              <th></th>
-            </tr>
-          </thead>
-          <tbody>
-            {state.map((event, index) => (<Event key={index} event={event} dispatch={dispatch}/>))}
-          </tbody>
-        </table>
+        <EventForm state={state} dispatch={dispatch}></EventForm>
+        <Events state={state} dispatch={dispatch}></Events>
       </div>
     </>
   );
